refactor: move SaunaConfig model to core app
