+++ backend/backend/apps/bookings/validators.py
@@ -3,7 +3,7 @@ import datetime
 from django.core.exceptions import ValidationError
 from django.utils.translation import gettext_lazy as _
 
-from .models import SaunaConfig
+from backend.apps.core.models import SaunaConfig
 
 sauna_config = SaunaConfig.get()
